chore: clean up quants
