+++ aphrodite/quantization/aqlm.py
@@ -11,7 +11,7 @@ import torch.nn.functional as F
 
 from aphrodite._C import ops
 from aphrodite.modeling.layers.linear import LinearMethodBase, set_weight_attrs
-from aphrodite.modeling.layers.quantization.base_config import (
+from aphrodite.quantization.base_config import (
     QuantizationConfig)
