+++ aphrodite/quantization/awq.py
@@ -8,7 +8,7 @@ from aphrodite.modeling.layers.fused_moe import (moe_align_block_size,
                                                  fused_moe, fused_topk)
 from aphrodite.modeling.layers.linear import (LinearMethodBase,
                                               set_weight_attrs)
-from aphrodite.modeling.layers.quantization.base_config import (
+from aphrodite.quantization.base_config import (
     QuantizationConfig)
