+++ aphrodite/quantization/bitsandbytes.py
@@ -5,7 +5,7 @@ from typing import List, Dict, Any, Optional, TypeVar, NamedTuple
 from aphrodite._C import ops
 from aphrodite.modeling.layers.linear import (LinearMethodBase,
                                               set_weight_attrs)
-from aphrodite.modeling.layers.quantization.base_config import (
+from aphrodite.quantization.base_config import (
     QuantizationConfig)
 from aphrodite.modeling.layers.linear import (ColumnParallelLinear,
                                               QKVParallelLinear,
