+++ aphrodite/quantization/eetq.py
@@ -5,7 +5,7 @@ import torch
 from torch.nn.parameter import Parameter
 
 from aphrodite.modeling.layers.linear import LinearMethodBase, set_weight_attrs
-from aphrodite.modeling.layers.quantization.base_config import \
+from aphrodite.quantization.base_config import \
     QuantizationConfig
 
 HAS_EETQ = False
