+++ aphrodite/quantization/quip.py
@@ -6,9 +6,9 @@ from torch.nn.parameter import Parameter
 from aphrodite._C import ops
 from aphrodite.modeling.layers.linear import (LinearMethodBase,
                                               set_weight_attrs)
-from aphrodite.modeling.layers.quantization.base_config import (
+from aphrodite.quantization.base_config import (
     QuantizationConfig)
-from aphrodite.modeling.layers.quantization.quip_utils import (
+from aphrodite.quantization.quip_utils import (
     get_packed_abs_grid,
     get_hadK,
     matmul_hadUt_cuda,
